Move stories to src/stories and MDX files to Storybook
