+++ src/stories/general.stories.jsx
@@ -2,19 +2,19 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text, select, boolean, array, number } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
-import Header from '../src/templates/Header';
-import NavBar from '../src/templates/NavBar';
-import Menu from '../src/components/Menu';
-import PageHeader from '../src/components/PageHeader';
+import Header from '../templates/Header';
+import NavBar from '../templates/NavBar';
+import Menu from '../components/Menu';
+import PageHeader from '../components/PageHeader';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import DataIcon from '../src/components/DataIcon';
-import Footer from '../src/templates/Footer';
-import ApiDocs from '../src/components/ApiDocs';
-import ToggleBlock from '../src/components/ToggleBlock';
-import ShowMoreContainer from '../src/components/ShowMoreContainer';
-import Announcement from '../src/templates/Announcement';
-import PublisherList from '../src/components/PublisherList';
-import "../src/theme/styles/index.scss";
+import DataIcon from '../components/DataIcon';
+import Footer from '../templates/Footer';
+import ApiDocs from '../components/ApiDocs';
+import ToggleBlock from '../components/ToggleBlock';
+import ShowMoreContainer from '../components/ShowMoreContainer';
+import Announcement from '../templates/Announcement';
+import PublisherList from '../components/PublisherList';
+import "../theme/styles/index.scss";
 
 import links from './data/menu.json';
 import orgs from './data/publishers.json';
