+++ src/stories/introduction.stories.mdx
@@ -1,7 +1,10 @@
----
-name: Home
-route: /
----
+import { Meta } from '@storybook/addon-docs';
+
+<Meta title="Introduction" />
+
+<style>{`
+
+`}</style>
 
 # Data Catalog Components
 A set of React components to facilitate the creation of Open Data Catalogs with React.
